Implement compound assignment operators
Adds +=, -=, *=, /=, &=, |=, ^=, <<=, >>= operators.
Tokeniser: 9 new token kinds; 3-char <<= and >>= scanned before 2-char.
Parser: assign_expr desugars a op= b to ND_ASSIGN(a, ND_BINOP(op, copy(a), b))
using a new copy_node() helper for AST duplication.
cpu3/cpu.py: fix div to use int() truncation (was leaving a float in r0).
13 new tests in tests/test_compound.sh; all 122 tests pass.

+++ cpu3/cpu.py
@@ -305,7 +305,7 @@ class CPU:
         elif    i == 'add':     s.r0 = m.read32(s.sp) + s.r0; s.sp += 4
         elif    i == 'sub':     s.r0 = m.read32(s.sp) - s.r0; s.sp += 4
         elif    i == 'mul':     s.r0 = m.read32(s.sp) * s.r0; s.sp += 4
-        elif    i == 'div':     s.r0 = m.read32(s.sp) / s.r0; s.sp += 4
+        elif    i == 'div':     s.r0 = int(m.read32(s.sp) / s.r0); s.sp += 4
         elif    i == 'mod':     s.r0 = m.read32(s.sp) % s.r0; s.sp += 4
         elif    i == 'shl':     s.r0 = m.read32(s.sp) << s.r0; s.sp += 4
         elif    i == 'shr':     s.r0 = m.read32(s.sp) >> s.r0; s.sp += 4
